TASK-459 Added new field to 'account.sdd.mandate' and changed an older one
The purpose of the new field id to replace 'creditor_id' in the form so that
only companies can be selected for 'creditor_id'. The latter was changed to a
'related' field that is computed directly from 'creditor_company_id' to ensure
that the code that depends on this model behaves the same way as before.